requirements.txt
click==7.1.2
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
future==0.18.2
numpy==1.19.1
opencv-python==4.3.0.36
Pillow==7.2.0
scipy==1.5.2
torch==1.6.0
torchvision==0.7.0
